requirements.txt
alt-profanity-check
argostranslate
